examine-growth: Have basic font stack
The problem is that Chrome for Windows’s default has a really ugly
rendering of Times New Roman unless we specify another font.  I
do not like using Georiga when there are a bunch of numbers, and I
do not like making the user download 120k or so of fonts, so we
have Cambria for Windows users, and Arial for most other users.