examine-growth: Fix bug caused by copypasta code
Since I did a copypasta for the code which converts a name with
quotes and accents in to a suitable filename, I had an issue with
the New Mexico page pointing to Doña Ana instead of the Donya Ana
version (which I need to use when mixing Windows native and Cygwin
programs in Windows 10, since they can not agree on the encoding to use
for accents in filenames).

There are only two accents in US country names: Doña Ana New Mexico and
Coös County New Hampshire.  Coös is spelled "Coos" in the CSV, so is a
non issue.  Doña Ana has the accent, so I need to make a workaround.