Move Arial before Helvetica in font stack
There is a bug in CentOS 8’s free “URW” version of Helvetica (make
that Nimbus Sans) where its vertical alignment is off in Firefox.

To work around this, we have to make Arial first (yes, there is a
workaround which allows any savvy Linux user to have Arial and all of the
other 1990s Microsoft Core Fonts, although Arial and Georgia are really
the only ones which are still usable here in 2020).  This may make things
less attractive on Macs, but considering how awful their “butterfly”
keyboards have been, how they are treating Epic games and how they are,
according to The Verge, are treating Wordpress, I am not too happy with
Apple right now.

https://www.theverge.com/2020/8/21/21396316/apple-wordpress-in-app-purchase-tax-update-store